final version of validator and product

+++ pixel_reader_validation/programs_projects/bitblt/main.cpp
@@ -37,8 +37,8 @@ int main(int argc, char **argv)
     {
         BitBlt(hdcCompatible, 0, 0, 1, 1, hdcScreen, 5, 5, SRCCOPY);
         int result = (int)bitPointer[2];
-        long timestamp = (long) (duration_cast<microseconds>(system_clock::now().time_since_epoch())).count();
-        printf("got color at:%ld:%d\n", timestamp, result);
+        long long timestamp = duration_cast<microseconds>(system_clock::now().time_since_epoch()).count();
+        printf("got color at:%lld:%d\n", timestamp, result);
         fflush(stdout);
     }
     return 0;
